feat(mochi): export pinGlobal; dedupe the site's Shiki highlighter per process
The site's `highlight.server.ts` initialised Shiki with eager, unguarded
top-level side effects (a `logger.warn` + a top-level-`await createShiki`) and
no per-process singleton guard, so every distinct module graph that evaluated
the file re-ran both — loading ~11 TextMate grammars each time. A `bun run build`
printed the "[highlight] shiki engine …" line four times (prebuild process, the
build's route-extraction import, and the two separate `Bun.build` compile
cohorts), and at production runtime every demo SSR bundle importing the module
allocated its own highlighter (~11 instances).

- Export the existing internal `pinGlobal` primitive from `mochi-framework` and
  add it to both `mochi-env` virtual-module templates (server re-export + client
  re-export; it is isomorphic — globalThis only), so it is usable from modules in
  the Svelte SSR graph.
- Wrap the highlighter's log + engine selection + `createShiki` + `highlightCode`
  in `pinGlobal(...)`, dropping the module-level top-level `await`. The log and
  the engine now initialise at most once per process.
- Move the highlight-free helpers `isDemoIndex`/`stripImageConfig` (and the
  `SourceSpec` type) into a new leaf `components/sourceUtils.ts` and import them
  there from `lib/docs.ts`, so the barrel-generating prebuild process no longer
  transitively pulls in the highlighter it never uses.

Build now logs the engine once; runtime collapses to a single highlighter per
process. Docs were never built repeatedly and doc markdown was already SSR-only —
this only removes the redundant engine initialisation.
