+++ packages/site/src/components/utils.ts
@@ -1,11 +1,13 @@
 import { highlightCode } from '../lib/highlight.server';
+import { isDemoIndex, stripImageConfig, type SourceSpec } from './sourceUtils';
+
+export { isDemoIndex, stripImageConfig, type SourceSpec } from './sourceUtils';
 
 export function delay(minMs: number, maxMs: number = minMs): Promise<void> {
   const ms = minMs + Math.random() * (maxMs - minMs);
   return new Promise((resolve) => setTimeout(resolve, ms));
 }
 
-export type SourceSpec = { label: string; path: string; lang?: string; showImageConfig?: boolean };
 type Source = { label: string; lang: string; html: string; styleHtml?: string };
 
 const cache = new Map<string, string>();
@@ -44,38 +46,6 @@ export async function loadSources(specs: SourceSpec[]): Promise<Source[]> {
       };
     }),
   );
-}
-
-export function isDemoIndex(path: string): boolean {
-  return path.endsWith('demoIndex.ts');
-}
-
-// Drop the `image: {…}` block (and its leading comment) — it's noise in every demo
-// except the image ones, which opt back in via `showImageConfig`.
-export function stripImageConfig(code: string): string {
-  const lines = code.split('\n');
-  const out: string[] = [];
-  let depth = 0;
-  let skipping = false;
-  for (const line of lines) {
-    if (skipping) {
-      depth += (line.match(/\{/g)?.length ?? 0) - (line.match(/\}/g)?.length ?? 0);
-      if (depth <= 0) {
-        skipping = false;
-      }
-      continue;
-    }
-    if (/^\s*image:\s*\{/.test(line)) {
-      while (out.length && /^\s*\/\//.test(out[out.length - 1]!)) {
-        out.pop();
-      }
-      depth = (line.match(/\{/g)?.length ?? 0) - (line.match(/\}/g)?.length ?? 0);
-      skipping = depth > 0;
-      continue;
-    }
-    out.push(line);
-  }
-  return out.join('\n');
 }
 
 const STYLE_RE = /\n<style(?:\s[^>]*)?>[\s\S]*?<\/style>\s*$/;
